moved files outside their folders for merge conflict resolution
