+++ src/components/Home/Home.tsx
@@ -1,7 +1,7 @@
 // eslint-disable-next-line no-use-before-define
 import React, { useContext } from 'react';
 import { Link } from 'react-router-dom';
-import ManualTestCreation from '../TestCreation/ManualTestContainer/ManualTestCreation';
+import ManualTestCreation from '../TestCreation/ManualTestCreation';
 import FileDirectory from '../FileDirectory/FileDirectory';
 import Monaco from '../Monaco/Monaco';
 // import * as monaco from 'monaco-editor'
